Add beat encoding to input and force output to stay in bounds
Two failures of current system are that it occasionally "gets lost"
about what beat it is on, and that it goes very high or very low.
First, this adds a beat encoding to the input to tell the network where
in the measure it is. Second, it uses a mask to set the probabilities
of playing outside the bounds to 0.